Move solvers to new solver module
