+++ pygfunction/solvers/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+from .base_solver import _BaseSolver
+from .detailed import Detailed
+from .similarities import Similarities
+from .equivalent import Equivalent
